Application for Mail_4_U(Version 2). To use, simply: 1. Enter name, email address & password 2. Enter # of emails you wish to reply to 3. Click 'run' to start application
Once you do, the app will start replying to the emails. Hopefully this reduces your workload, and if you have any
suggestions, do say so. Ty and GB! 😀😀😀

P.S: Try not to copy anything while app is replying to emails, as it might mess up the application. Thank you!